Update README.md
This is my personal project. It aims to provide people with a powerful tool for their studies — including schedules, a question bank, and a store for selling e-books to help support the platform’s costs.

The platform follows a freemium model and continues to grow. It's an ambitious project that I'm passionate about, and I’ve been studying hard to develop the best version of it.

Because of this focus, I currently have a limited number of public repositories — my priority has been improving my skills to build this project effectively.